zerofill: Change the type of len argument of glfs_zerofill() to off_t
glfs_zerofill() can be potentially called to zero-out entire file and
hence allow for bigger value of length parameter.

BUG: 1028673
Reviewed-on: http://review.gluster.org/6266
Tested-by: Gluster Build System <jenkins@build.gluster.com>
Tested-by: M. Mohan Kumar <mohan@in.ibm.com>

+++ api/src/glfs-fops.c
@@ -1877,7 +1877,7 @@ glfs_discard_async (struct glfs_fd *glfd, off_t offset, size_t len,
 }
 
 int
-glfs_zerofill_async (struct glfs_fd *glfd, off_t offset, size_t len,
+glfs_zerofill_async (struct glfs_fd *glfd, off_t offset, off_t len,
                       glfs_io_cbk fn, void *data)
 {
         struct glfs_io *gio  = NULL;
@@ -2927,7 +2927,7 @@ out:
 }
 
 int
-glfs_zerofill (struct glfs_fd *glfd, off_t offset, size_t len)
+glfs_zerofill (struct glfs_fd *glfd, off_t offset, off_t len)
 {
         int               ret             = -1;
         xlator_t         *subvol          = NULL;
