+++ xlators/cluster/dht/src/dht-common.h
@@ -696,7 +696,7 @@ int32_t dht_fallocate(call_frame_t *frame, xlator_t *this, fd_t *fd,
 int32_t dht_discard(call_frame_t *frame, xlator_t *this, fd_t *fd,
 		    off_t offset, size_t len, dict_t *xdata);
 int32_t dht_zerofill(call_frame_t *frame, xlator_t *this, fd_t *fd,
-                    off_t offset, size_t len, dict_t *xdata);
+                    off_t offset, off_t len, dict_t *xdata);
 
 int32_t dht_init (xlator_t *this);
 void    dht_fini (xlator_t *this);
